requirements.txt
accelerate>=0.33.0
datasets>=2.20.0
evaluate>=0.4.2
matplotlib>=3.9.0
numpy>=1.26.4
pandas>=2.2.2
pyyaml>=6.0.2
rouge-score>=0.1.2
scikit-learn>=1.5.1
sentence-transformers>=3.0.1
streamlit>=1.37.0
torch>=2.2.2
tqdm>=4.66.4
transformers>=4.44.0
